+++ dojo/system_settings/__init__.py
@@ -1,0 +1,1 @@
+import dojo.system_settings.admin  # noqa: F401
