+++ dojo/system_settings/admin.py
@@ -1,0 +1,5 @@
+from django.contrib import admin
+
+from dojo.system_settings.models import System_Settings
+
+admin.site.register(System_Settings)
