+++ dojo/system_settings/api/__init__.py
@@ -1,0 +1,1 @@
+path = "system_settings"  # noqa: RUF067
