+++ dojo/urls.py
@@ -37,7 +37,6 @@ from dojo.api_v2.views import (
     SLAConfigurationViewset,
     SonarqubeIssueTransitionViewSet,
     SonarqubeIssueViewSet,
-    SystemSettingsViewSet,
     ToolConfigurationsViewSet,
     ToolProductSettingsViewSet,
     ToolTypesViewSet,
@@ -75,7 +74,8 @@ from dojo.reports.urls import urlpatterns as reports_urls
 from dojo.search.urls import urlpatterns as search_urls
 from dojo.sla_config.urls import urlpatterns as sla_urls
 from dojo.survey.urls import urlpatterns as survey_urls
-from dojo.system_settings.urls import urlpatterns as system_settings_urls
+from dojo.system_settings.api.urls import add_system_settings_urls
+from dojo.system_settings.ui.urls import urlpatterns as system_settings_urls
 from dojo.test.api.urls import add_test_urls
 from dojo.test.ui.urls import urlpatterns as test_urls
 from dojo.test_type.urls import urlpatterns as test_type_urls
@@ -136,7 +136,7 @@ v2_api.register(r"risk_acceptance", RiskAcceptanceViewSet, basename="risk_accept
 v2_api.register(r"sla_configurations", SLAConfigurationViewset, basename="sla_configurations")
 v2_api.register(r"sonarqube_issues", SonarqubeIssueViewSet, basename="sonarqube_issue")
 v2_api.register(r"sonarqube_transitions", SonarqubeIssueTransitionViewSet, basename="sonarqube_issue_transition")
-v2_api.register(r"system_settings", SystemSettingsViewSet, basename="system_settings")
+v2_api = add_system_settings_urls(v2_api)
 v2_api.register(r"technologies", AppAnalysisViewSet, basename="app_analysis")
 v2_api = add_test_urls(v2_api)
 v2_api.register(r"tool_configurations", ToolConfigurationsViewSet, basename="tool_configuration")
